Start work on mesh scheduling + fix up LocalMeshObject
